+++ docx/oxml/__init__.py
@@ -246,3 +246,19 @@ from .text.run import CT_Br, CT_R, CT_Text  # noqa
 register_element_cls('w:br', CT_Br)
 register_element_cls('w:r',  CT_R)
 register_element_cls('w:t',  CT_Text)
+
+from .text_box import CT_AlternateContent, CT_AC_Choice, CT_AC_Fallback, CT_Anchor, CT_Drawing, CT_PositionH, \
+    CT_PositionV, CT_Pick, CT_Group, CT_Shape, CT_Textbox, CT_TxbxContent  # noqa
+
+register_element_cls('mc:AlternateContent', CT_AlternateContent)
+register_element_cls('mc:Choice', CT_AC_Choice)
+register_element_cls('w:drawing', CT_Drawing)
+register_element_cls('wp:anchor', CT_Anchor)
+register_element_cls('wp:positionH', CT_PositionH)
+register_element_cls('wp:positionV', CT_PositionV)
+register_element_cls('mc:Fallback', CT_AC_Fallback)
+register_element_cls('w:pict', CT_Pick)
+register_element_cls('v:group', CT_Group)
+register_element_cls('v:shape', CT_Shape)
+register_element_cls('v:textbox', CT_Textbox)
+register_element_cls('v:txbxContent', CT_TxbxContent)